requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
pyodbc
waitress
opencv-python
pathlib
pyzbar
openpyxl
python-barcode
pyodbc
pywin32
Pillow